Renamed `SessionController` to `GoogleOAuthHandler`
